// cpo-server/src/types/Setting.ts
export interface AssetSetting {
  connections: AssetConnectionSetting[];
}

export interface AssetConnectionSetting {
  id: string;
  name: string;
  description: string;
  url: string;
  timestamp: Date;
  type: AssetConnectionType;
  refreshIntervalMins?: number;
  schneiderConnection?: AssetSchneiderConnectionType;
  greencomConnection?: AssetGreencomConnectionType;
  iothinkConnection?: AssetIothinkConnectionType;
  witConnection?: AssetWitConnectionType;
  lacroixConnection?: AssetLacroixConnectionType;
  schneiderTestConnection?: AssetSchniederTestConnetionType;
}

export enum AssetConnectionType {
  SCHNEIDER_TEST ='schneider_test',
  SCHNEIDER = 'schneider',
  GREENCOM = 'greencom',
  IOTHINK = 'iothink',
  WIT = 'wit',
  LACROIX = 'lacroix'
}

export interface AssetUserPasswordConnectionType {
  user: string;
  password: string;
}

export interface AssetTokenConnectionType {
  token: string;
  site_id: string;
}

export interface AssetGreencomConnectionType {
  clientId: string;
  clientSecret: string;
}

export type AssetSchneiderConnectionType = AssetUserPasswordConnectionType;

export type AssetSchniederTestConnetionType = AssetTokenConnectionType;

export type AssetIothinkConnectionType = AssetUserPasswordConnectionType;

export type AssetLacroixConnectionType = AssetUserPasswordConnectionType;

export interface AssetWitConnectionType extends AssetUserPasswordConnectionType {
  clientId: string;
  clientSecret: string;
  authenticationUrl: string;
}